Fix some use-after-free errors
Since `underlying` is computed on the fly now, if it's not stored by the
caller, it's dropped immediately. So if you try to get the underlying
and immediately cast it to something, it will destruct before you have
a chance to make use of the downcast value.

This is super annoying, and I'm glad it came up in our test case, but I
would love to figure out how to make it a compile error.